Set media width globally for tests
